Disable GCC's stack smashing protection
Apparently, GCC generates code that uses one of its own built-in
libraries, libssp, when it compiles functions that contain arrays
allocated on the stack.

Since liblinux and its examples are designed to be freestanding,
they do not link to GCC's built-in stack protection library.
This results in errors such as:

    undefined reference to `__stack_chk_fail'

Disabling the stack protector avoids implicitly depending on libssp.
The -fno-stack-protector flag is used to disable the security feature.
It is documented in the following page:

    https://gcc.gnu.org/onlinedocs/gcc/Instrumentation-Options.html