Introduce liblinux_start process startup function
The _start function passes the stack pointer to liblinux_start,
which figures out the location of the arguments and environment
and then calls the main function.

There doesn't seem to be a good way to write _start in C.
GCC assumes _start may be called by other functions and have return
addresses pushed on the stack, so it generates code to realign rsp.
It does this even if _start is declared with __attribute__((noreturn))
or ends with __builtin_unreachable().
This makes it hard to obtain access to the data allocated on the stack.
It is easier to pass the stack pointer to a C function from assembly
then work out the addresses in C.