Declare auxiliary value structure
The kernel provides all new processes with a vector of auxiliary data.
Entries are structures that pair two numbers: type and value.

The type specifies how to interpret the value.
It encodes not just the data type of the value but also its semantics.
For example, the AT_RANDOM type specifies two things:

 - The value is a pointer
 - It points to a 16-byte array filled with random data