Automatically generate build rules for objects
When multiple rules for the same target are specified,
their prerequisites are merged into one big dependency list.
Some compilers can output make rules that take advantage of this:
the compiler's output is merged with rules written by the programmer.

The current build system implementation uses pattern rules
in order to match object files to their sources.
This is incompatible with compiler output, so it must be changed.

The new implementation generates rules for object files
metaprogrammatically.