Define gcc_dependency_generation_options function
This make function generates a set of options that instruct GCC to generate
dependency information from an input file, add extra targets to it and then
write the data to an output file.

This happens separately from the compilation step since support for
alternative compilers is desirable. This way, supported compilers can be
used to generate dependency information and other compilers can be used
to actually compile the code.
In any case, these options should also be compatible with clang.

References:

https://gcc.gnu.org/onlinedocs/gcc/Preprocessor-Options.html
https://clang.llvm.org/docs/ClangCommandLineReference.html#dependency-file-generation
https://github.com/matheusmoreira/liblinux/issues/1#issuecomment-472703122