Define generate_dependency_file_rule function
This make function generates rules for creating dependency data files.
The 1st argument is the target, the dependency data file to be created.
The 2nd argument is the source whose dependencies will be computed.
The 3rd argument is the name of a make function which will be applied
to the 2nd argument.
The function must produce the list of targets that depend on the source,
and therefore also depend on its dependencies.