Remove trailing slash from build directory pattern
According to git documentation:

    If the pattern ends with a slash, it is removed
    for the purpose of the following description,
    but it would only find a match with a directory.
    In other words, foo/ will match a directory foo
    and paths underneath it,
    but will not match a regular file or a symbolic link foo.

Removing the trailing slash makes it possible for the build directory
to be a symbolic link to another directory.

This enables a simple way to build liblinux in memory:

    $ mkdir -p /dev/shm/liblinux/build
    $ ln -s /dev/shm/liblinux/build build

Adding a trailing slash to the parameter of rm prevents the command from
deleting the symbolic link itself.

References:

https://git-scm.com/docs/gitignore