Refactor object rule generation functions
Now there is a rule template, multiple useful functions and documentation.

The new implementation is much easier to understand
since it makes use of existing path conversion functions.
It operates only on the source file path,
computing all other paths as needed.

The new template also includes the source file's dependency data file
among the object file's prerequisites.
This triggers the generation of dependency data when the object is updated.

The variables are no longer undefined after they are used.

Related to issue #8.