Refactor dependency file rule generator functions
Now there is a rule template, a function that evaluates the template
and more detailed documentation.

The new implementation is much easier to understand
since it makes use of a new path conversion function.
It operates only on the source file path:
the path to the target is computed from it.

The variables are no longer undefined after they are used.