Create build directory pointed to by symbolic link
The build directory may be a symbolic link to another location.
This allows building in locations such as /dev/shm.
However, there is a bug: if the directory pointed to
by the symbolic link does not exist, make will fail.

Now the build system ensures the existence of the target directory
if the build directory is a symbolic link.

Closes #12.