Add r8, r9 and r10 to the list of outputs
On Linux x86_64, system calls clobbers rcx, r8, r9, r10 and r11.
However, r8, r9 and r10 are inputs to the syscall instruction.
The GCC manual states:

> When the compiler selects the registers to use
> to represent the input operands,
> it does not use any of the clobbered registers.

So they can't be placed in the list of clobbered registers.
Placing them in the list of outputs achieves the same result,
even though they are not true outputs.

References:

https://gcc.gnu.org/onlinedocs/gcc/Extended-Asm.html#Input-Operands
https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/tools/include/nolibc/nolibc.h