Use the "+" prefix for the rax register constraint
The rax register is simultaneously an input and output register
since it Linux reads the system call number from it
and later writes the return value to it.
The "+" prefix was designed for these situations:

> Operands using the ‘+’ constraint modifier
> count as two operands (that is, both as input and output)

This allows rax to be listed only once, avoiding needless repetition.
The list of inputs now contains just the six system call parameters.
All operand register constraints are now generic
and only the constraints applied to the variables they reference matter.
The return_value variable is not needed anymore and can be removed.

References:

https://gcc.gnu.org/onlinedocs/gcc/Extended-Asm.html#OutputOperands