Use the "+" constraint prefix for r8, r9 and r10
The r8, r9 and r10 registers are system call inputs
that may be clobbered by the Linux kernel.
Since the compiler won't use clobbered registers as inputs or outputs,
they must be placed in the outputs list to get the same effect.

The "+" output register constraint prefix can be used instead of "="
for output registers that are also inputs. With this notation,
there is no need for registers to be in both input and output lists
simultaneously.

Now every register is listed exactly once.

References:

https://gcc.gnu.org/onlinedocs/gcc/Extended-Asm.html#OutputOperands

> Output constraints must begin with either ‘=’
> (a variable overwriting an existing value)
> or ‘+’ (when reading and writing).

> Operands using the ‘+’ constraint modifier
> count as two operands (that is, both as input and output)

> When the compiler selects the registers to use
> to represent the output operands,
> it does not use any of the clobbered registers.

https://gcc.gnu.org/onlinedocs/gcc/Extended-Asm.html#InputOperands

> When the compiler selects the registers to use
> to represent the input operands,
> it does not use any of the clobbered registers.

https://gcc.gnu.org/onlinedocs/gcc/Extended-Asm.html#Clobbers-and-Scratch-Registers

> There is no way to specify that input operands get modified
> without also specifying them as output operands.

> When the compiler selects which registers to use to represent input and output operands,
> it does not use any of the clobbered registers.