Strip the results of the find make function
When find was invoked with a predicate function,
the result was the empty string for every path
that did not satisfy the predicate function.
This resulted in a lot of extraneous white spaces
in the positions the paths would have otherwise occupied.
Stripping the result gets rid of the spaces.

The recurse and file_system.traverse functions didn't need this fix
because they always return elements in each iteration of foreach.
There are no spaces to strip out.