Switch to configuration-based builds
Gets rid of the messy static/dynamic variables and functions.
The build tree is now rooted on the build configuration
which consists of target architecture and linkage type.
The build system will now configure itself for a static or dynamic build
instead of building both at the same time.

    linkage=static  make    # Builds statically linked targets
    linkage=dynamic make    # Builds dynamically linked targets

This simplifies the build tree and the build system code
and allows more configurations to be added in the future.