Fix buggy implementation of the equal? function
Make's subst function replaces all occurrences of the string it finds.
This causes the equal? function to return an incorrect result
when the second string consists of multiple copies of the first:

    $(call equal?,a,aa)    # =
    $(call equal?,aa,a)    # ≠

The correct implementation returns true if both substitutions are empty.

Asked about this on Stack Overflow 3 years ago,
got an answer by the author of the GNU Make Standard Library
and then completely forgot about it until today.

References:

    https://www.gnu.org/software/make/manual/html_node/Text-Functions.html
    https://stackoverflow.com/q/37469553/512904
    https://stackoverflow.com/a/37479819/512904