Refactor the examples system to be more general
Usage examples are simply executables that are linked to a library.
The build tree should have an executables directory instead.
The directory that contains the executable's source is reflected there
and denotes what type of executable it is.

Now all the path conversion functions don't use any project-specific
structural variables.