Set .DEFAULT_GOAL in the generated project rules
Since lots of rules are generated and their order is undefined,
it is wise to set make's default target explicitly.

Normally, the default goal will be the first target of the project.
In the following example project:

    example.targets := X Y Z

The default target is X. This mirrors make's default behavior.

The default target can also be specified explicitly:

    example.targets.default := Z