Pass the result of function calls to the templates
The generated rules should be as simple and explicit as possible.
They should not reference variables or functions from the framework.
If the function call is escaped, it is passed as code to the template.
This results in references to framework functions in the generated code.
Simply passing the results of the functions as arguments avoids this.