Define the resolve.first make function
This function computes the value of the first defined variable
in a list of namespaces.

Should be useful when values set in specific contexts
override values specified in more general contexts.
For example, a project might want to use standard C99 for all targets
except for a specific file which must be compiled with GNU extensions.